Prevent invalid dates
If a user did not enter a value in one of the three boxes or entered an invalid value, the date function would return December 31, 1969. POST values need to be checked to confirm they are valid before they are formatted and inserted into the email. If the value as a whole is invalid, proceed as if they entered nothing.